Keep package.json on bower install.
When running Jest tests, Scout Asia rely on a components
`package.json` to resolve Origami components which where installed via Bower.

In v3 "package.json" was added to Bowers ignore property, which
means their tests now fail. The recieved v3 without taking action
because it's included in other components such as o-tooltip using a
range https://github.com/Financial-Times/o-tooltip/pull/71. We use
a range as the interface of v3 is the same as v2 -- we just removed
the dom-delegate alias.